requirements.txt
fastapi==0.115.12
google-ai-generativelanguage==0.6.15
google-api-python-client==2.169.0
google-generativeai==0.8.5
langchain==0.3.23
numpy==2.2.4
openai==0.28.0
pydantic==2.11.3
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
SQLAlchemy==2.0.40
uvicorn==0.34.1
